Log the state as an integer, not as a byte with value 0, 1 or 2
It is rather hard to see in a terminal window whether a char with value 0, 1
or 2 has been printed;)

The elegant thing to do would of course be to make the enum in question into
an enum class, and define an operator<< for it, to output it symbolically. But
I can not be distracted into that now.

+++ sc/source/core/data/formulacell.cxx
@@ -3687,7 +3687,7 @@ bool ScFormulaCell::InterpretFormulaGroup()
     mxGroup->meCalcState = sc::GroupCalcRunning;
     if (!sc::FormulaGroupInterpreter::getStatic()->interpret(*pDocument, mxGroup->mpTopCell->aPos, mxGroup, aCode))
     {
-        SAL_INFO("sc.opencl", "interpreting group " << mxGroup << " (state " << mxGroup->meCalcState << ") failed, disabling");
+        SAL_INFO("sc.opencl", "interpreting group " << mxGroup << " (state " << (int) mxGroup->meCalcState << ") failed, disabling");
         mxGroup->meCalcState = sc::GroupCalcDisabled;
         return false;
     }
